+++ src/main.js
@@ -1,8 +1,11 @@
 import { LitElement, html } from 'lit-element';
 import { classMap } from 'lit-html/directives/class-map';
+import { styleMap } from 'lit-html/directives/style-map';
 import ResizeObserver from 'resize-observer-polyfill';
 import MediaPlayerObject from './model';
 import style from './style';
+import sharedStyle from './sharedStyle';
+import handleClick from './utils/handleClick';
 
 import './components/groupList';
 import './components/dropdown';
@@ -20,16 +23,9 @@ import {
   LABEL_SHORTCUT,
 } from './const';
 
-if (!customElements.get('mwc-button')) {
+if (!customElements.get('ha-slider')) {
   customElements.define(
-    'mwc-button',
-    class extends customElements.get('paper-button') {},
-  );
-}
-
-if (!customElements.get('ha-paper-slider')) {
-  customElements.define(
-    'ha-paper-slider',
+    'ha-slider',
     class extends customElements.get('paper-slider') {},
   );
 }
@@ -63,7 +59,10 @@ class MiniMediaPlayer extends LitElement {
   }
 
   static get styles() {
-    return style;
+    return [
+      sharedStyle,
+      style,
+    ];
   }
 
   set hass(hass) {
@@ -106,6 +105,9 @@ class MiniMediaPlayer extends LitElement {
       source: 'default',
       sound_mode: 'default',
       toggle_power: true,
+      tap_action: {
+        action: 'more-info',
+      },
       ...config,
       hide: { ...DEFAULT_HIDE, ...config.hide },
       speaker_group: {
@@ -166,8 +168,9 @@ class MiniMediaPlayer extends LitElement {
 
     return html`
       <ha-card
-        @click=${this.handleMoreInfo}
         class=${this.computeClasses()}
+        style=${this.computeStyles()}
+        @click=${e => this.handlePopup(e)}
         artwork=${config.artwork}
         content=${this.player.content}>
         <div class='mmp__bg'>
@@ -194,7 +197,8 @@ class MiniMediaPlayer extends LitElement {
             ${!config.collapse && this.player.active ? html`
               <mmp-media-controls
                 .player=${this.player}
-                .config=${config}>
+                .config=${config}
+                .break=${this.break}>
               </mmp-media-controls>
             ` : ''}
             <mmp-shortcuts
@@ -216,10 +220,12 @@ class MiniMediaPlayer extends LitElement {
           </div>
         </div>
         <div class='mmp__container'>
-          <mmp-progress
-            .player=${this.player}
-            .showTime=${!this.config.hide.runtime}>
-          </mmp-progress>
+          ${this.player.active && this.player.hasProgress ? html`
+            <mmp-progress
+              .player=${this.player}
+              .showTime=${!this.config.hide.runtime}>
+            </mmp-progress>
+          ` : ''}
         </div>
       </ha-card>
     `;
@@ -237,6 +243,11 @@ class MiniMediaPlayer extends LitElement {
     return html`<div class='cover' style='background-image: ${url};'></div>`;
   }
 
+  handlePopup(e) {
+    e.stopPropagation();
+    handleClick(this, this._hass, this.config, this.config.tap_action, this.player.id);
+  }
+
   renderIcon(artwork) {
     if (this.config.hide.icon) return;
     if (this.player.active && artwork && this.config.artwork === 'default')
@@ -247,9 +258,10 @@ class MiniMediaPlayer extends LitElement {
           state=${this.player.state}>
         </div>`;
 
+    const state = !this.config.hide.icon_state && this.player.isActive;
     return html`
-      <div class='entity__icon'>
-        <ha-icon .icon=${this.computeIcon()}></ha-icon>
+      <div class='entity__icon' ?color=${state}>
+        <ha-icon .icon=${this.computeIcon()} ></ha-icon>
       </div>`;
   }
 
@@ -293,12 +305,21 @@ class MiniMediaPlayer extends LitElement {
       '--initial': this.initial,
       '--bg': config.background,
       '--group': config.group,
-      '--more-info': config.more_info,
+      '--more-info': config.tap_action !== 'none',
       '--has-artwork': this.player.hasArtwork && this.thumbnail,
       '--flow': config.flow,
       '--collapse': config.collapse,
       '--rtl': this.rtl,
+      '--progress': this.player.hasProgress,
       '--runtime': !config.hide.runtime && this.player.hasProgress,
+      '--inactive': !this.player.isActive,
+    });
+  }
+
+  computeStyles() {
+    const { scale } = this.config;
+    return styleMap({
+      ...(scale && { '--mmp-unit': `${40 * scale}px` }),
     });
   }
 
@@ -344,11 +365,17 @@ class MiniMediaPlayer extends LitElement {
     this.edit = !this.edit;
   }
 
-  handleMoreInfo() {
-    if (!this.config.more_info) return;
-    const e = new Event('hass-more-info', { composed: true });
-    e.detail = { entityId: this.config.entity };
+  fire(type, inDetail, inOptions) {
+    const options = inOptions || {};
+    const detail = inDetail === null || inDetail === undefined ? {} : inDetail;
+    const e = new Event(type, {
+      bubbles: options.bubbles === undefined ? true : options.bubbles,
+      cancelable: Boolean(options.cancelable),
+      composed: options.composed === undefined ? true : options.composed,
+    });
+    e.detail = detail;
     this.dispatchEvent(e);
+    return e;
   }
 
   updateIdleStatus() {